Support refreshing statistics when car is unplugged
When the car is unplugged, the `active` API endpoint returns an HTTP
404 status which causes the `v2hDevice.refresh_stats` method to
error. However, the majority of the stats are still available and
valid so `refresh_stats` should still return the available stats
without error.

This change simply wraps the call to the `active` endpoint in a
try/except block and sets the active attibrute to an empty dict; this
is the same result as if the API call had not been made in the first
place and so seems the most appropriate placeholder value.

Fixes #16